move logic into data layer
