+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2024-05-08T09:54:22.766Z\n"
-"PO-Revision-Date: 2024-05-08T09:54:22.767Z\n"
+"POT-Creation-Date: 2024-06-10T12:10:57.544Z\n"
+"PO-Revision-Date: 2024-06-10T12:10:57.544Z\n"
 
 msgid ""
 "THIS NEW RELEASE INCLUDES SHARING SETTINGS PER INSTANCES. FOR THIS VERSION "
